solved issues in user auth and login system

+++ Project/views.py
@@ -10,43 +10,53 @@ from django.template import RequestContext
 from django.core.files.storage import FileSystemStorage
 from django.db.models import *
 from ajaxuploader.views import AjaxFileUploader
+from django.contrib.auth.decorators import login_required
+from django.views.decorators.cache import cache_control
+from django.template.response import TemplateResponse
+from django.contrib.postgres.search import SearchVector
 BASE_DIR = os.path.dirname(os.path.dirname(os.path.abspath(__file__)))
 
 # Create your views here.
-def index(request):
+
+
+@cache_control(no_cache=True, must_revalidate=True, no_store=True)
+def index(request):   
+      
+    if request.session.get('0',False) is False :return HttpResponseRedirect('/users_auth/login/') 
     rates=Project_User_Donation.objects.values('prj_id').annotate(Sum('rate'))
     pics=Project_pics.objects.all()
-    projects=Projects.objects.all()
+    projects=Projects.objects.all().order_by("-updated_at")
     return render(request, 'Project/projects.html',{'projects':projects,"rates":rates,"pics":pics})
-
+@cache_control(no_cache=True, must_revalidate=True, no_store=True)
 def categories(request):
     
     categories = Categories.objects.all()
     return render(request, 'users_auth/categories.html',{'categories':categories})
-
+@cache_control(no_cache=True, must_revalidate=True, no_store=True)
 def addcategory(request):
     if request.method == 'POST':
         category = Categories.objects.create(title=request.POST.get("catName", ""))
         category.save()
     return categories(request)
-
+@cache_control(no_cache=True, must_revalidate=True, no_store=True)
 def addproject(request):
-    
+    if request.session.get('0',False) is False :return HttpResponseRedirect('/users_auth/login/') 
     if request.method == 'POST':
-       details = ProjectForm(request.POST)
+       details = ProjectForm(request.POST,user_id=request.session.get('0'))
        if details.is_valid():   
   
             # Temporarily make an object to be add some 
             # logic into the data if there is such a need 
             # before writing to the database    
             project = details.save(commit = False) 
-  
+            project.user_id = request.session.get('0')
             # Finally write the changes into database 
-            project.save()   
+            project.save() 
+              
         
             # redirect it to some another page indicating data 
             # was inserted successfully 
-            return HttpResponse("data submitted successfully") 
+            return HttpResponseRedirect('/project/')
               
        else: 
           
@@ -62,15 +72,13 @@ def addproject(request):
         form = ProjectForm(None)
         categories = Categories.objects.all()    
         return render(request, 'Project/add_project.html', {'form':form,'categories':categories}) 
-
+@cache_control(no_cache=True, must_revalidate=True, no_store=True)
 def django_image_and_file_upload_ajax(request):
+    if request.session.get('0',False) is False :return HttpResponseRedirect('/users_auth/login/') 
     if request.method == 'POST':
        form = ImageFileUploadForm(request.POST, request.FILES)
        if form.is_valid():
-        #myfile = request.FILES['picture']
-        #fs = FileSystemStorage(location='D:/ITI/python/Project/fund/media/'+str(prj_id))
-        #filename = fs.save(myfile.name, myfile)
-        #uploaded_file_url = fs.url(filename)
+
         tmp=Projects.objects.get(id=request.POST.get('prj_pic'))
         pic=request.FILES['picture']
         request.FILES['picture'].name = str(request.POST.get('prj_pic'))+"_"+request.FILES['picture'].name
@@ -82,8 +90,11 @@ def django_image_and_file_upload_ajax(request):
            return JsonResponse({'error': True, 'errors': form.errors})
     else:
         form = ImageFileUploadForm()
-        return render(request, 'Project/project_images.html', {'form': form})
+        projects = Users.objects.filter(id=int(request.session.get('0')))[0].users.all()
+        return render(request, 'Project/project_images.html', {'form': form,'projects':projects})
+@cache_control(no_cache=True, must_revalidate=True, no_store=True)
 def project(request,prj_id):
+    if request.session.get('0',False) is False :return HttpResponseRedirect('/users_auth/login/') 
     if request.method == 'POST':
         project = Projects.objects.get(id=prj_id)
         project.Nor=project.Nor+1
@@ -100,45 +111,56 @@ def project(request,prj_id):
             pics = project.oproject.all()
             comments=Project_comments.objects.filter(prj_comment=prj_id).order_by('updated_at').reverse()
             users=Users.objects.all()
+            tags=project.tags.split(",")
+            if tags:
+                if len(tags)==2:
+                    similar=Projects.objects.filter(Q(tags__contains=tags[0]) | Q(tags__contains=tags[1]))
+                else:
+                    similar=Projects.objects.filter(tags__contains=tags[0])
             if(project.Nor!=0):
                 overall=project.rate/project.Nor
             else:
                 overall=0
             if rates :
-                return render(request, 'Project/project.html', {'project': project,"pics":pics,"overall":overall,"rates":rates[0]['rate__sum'],"comments":comments,"users":users})
+                return render(request, 'Project/project.html', {'similar':similar,'project': project,"pics":pics,"overall":overall,"rates":rates[0]['rate__sum'],"comments":comments,"users":users})
             else :
-                return render(request, 'Project/project.html', {'project': project,"pics":pics,"overall":overall,"comments":comments,"users":users})
+                return render(request, 'Project/project.html', {'similar':similar,'project': project,"pics":pics,"overall":overall,"comments":comments,"users":users})
         else:
                     return HttpResponseRedirect('/project/')
 
-
+@cache_control(no_cache=True, must_revalidate=True, no_store=True)
 def addcomment(request,prj_id):
+    if request.session.get('0',False) is False :return HttpResponseRedirect('/users_auth/login/')
     if request.method == 'POST':
         comment = Project_comments.objects.create(title=request.POST.get('title'),prj_comment=Projects.objects.get(id=prj_id),user=Users.objects.get(id=1)) 
         comment.save()
         return HttpResponseRedirect('/project/'+str(prj_id)+'/details')
-        
+@cache_control(no_cache=True, must_revalidate=True, no_store=True)
 def addlike(request,prj_id):
+    if request.session.get('0',False) is False :return HttpResponseRedirect('/users_auth/login/')
     if request.method == 'POST':
         comment = Project_comments.objects.get(id=request.POST.get('comment_id'))
         comment.likes = comment.likes + 1
         comment.save()
         return HttpResponseRedirect('/project/'+str(prj_id)+'/details')
-
+@cache_control(no_cache=True, must_revalidate=True, no_store=True)
 def adddislike(request,prj_id):
+    if request.session.get('0',False) is False :return HttpResponseRedirect('/users_auth/login/')
     if request.method == 'POST':
         comment = Project_comments.objects.get(id=request.POST.get('comment_id'))
         comment.dislikes = comment.dislikes + 1
         comment.save()
         return HttpResponseRedirect('/project/'+str(prj_id)+'/details')
-    
+@cache_control(no_cache=True, must_revalidate=True, no_store=True)
 def deletecomment(request,prj_id):
+    if request.session.get('0',False) is False :return HttpResponseRedirect('/users_auth/login/')
     if request.method == 'POST':
         Project_comments.objects.get(id=request.POST.get('comment_id')).delete()
         return HttpResponseRedirect('/project/'+str(prj_id)+'/details')
     
-
+@cache_control(no_cache=True, must_revalidate=True, no_store=True)
 def donate(request,prj_id):
+    if request.session.get('0',False) is False :return HttpResponseRedirect('/users_auth/login/')
     if request.method == 'POST':
         rates=Project_User_Donation.objects.values('prj_id').annotate(Sum('rate')).filter(prj_id=prj_id)
         project=Projects.objects.get(id=prj_id)
@@ -150,15 +172,17 @@ def donate(request,prj_id):
             donation = Project_User_Donation.objects.create(prj=Projects.objects.get(id=prj_id),user=Users.objects.get(id=1),rate=request.POST.get('donation_amount')) 
             donation.save()
         return HttpResponseRedirect('/project/'+str(prj_id)+'/details')
-    
+@cache_control(no_cache=True, must_revalidate=True, no_store=True)
 def addreport(request,prj_id):
+    if request.session.get('0',False) is False :return HttpResponseRedirect('/users_auth/login/')
     if request.method == 'POST':
         project = Projects.objects.get(id=prj_id)
         project.reports = project.reports + 1
         project.save()
         return HttpResponseRedirect('/project/'+str(prj_id)+'/details')
-
+@cache_control(no_cache=True, must_revalidate=True, no_store=True)
 def deleteproject(request,prj_id):
+    if request.session.get('0',False) is False :return HttpResponseRedirect('/users_auth/login/')   
     if request.method == 'POST':
         project=Projects.objects.get(id=prj_id)
         rates=Project_User_Donation.objects.values('prj_id').annotate(Sum('rate')).filter(prj_id=prj_id)
@@ -168,6 +192,15 @@ def deleteproject(request,prj_id):
                 project.delete()
         else: 
             project.delete()
-        return HttpResponseRedirect('/project/'+str(prj_id)+'/details')
+        return HttpResponseRedirect('/project/')
 
-    
+@cache_control(no_cache=True, must_revalidate=True, no_store=True)
+def logout(request):
+    request.session[0]=False
+    return HttpResponseRedirect('/users_auth/login/')
+
+def person(request):
+    try :
+        return {'person': Users.objects.filter(id= request.session.get('0'))[0]} # of course some filter here
+    except:
+        return {}
